38-redux-UI-call

+++ client/src/components/message/RightSide.jsx
@@ -15,6 +15,7 @@ import {
   loadMoreMessages,
 } from '../../redux/actions/messageAction';
 import LoadIcon from '../../images/loading.gif';
+
 //
 const RightSide = () => {
   const { auth, message, theme, socket } = useSelector((state) => state);
@@ -153,8 +154,32 @@ const RightSide = () => {
   }, [isLoadMore]);
 
   const handleDelteConversation = () => {
-    dispatch(deleteConversation({ auth, id }));
-    return history.push('/message');
+    if (window.confirm('Do you want to delete?')) {
+      dispatch(deleteConversation({ auth, id }));
+      return history.push('/message');
+    }
+  };
+
+  // Call
+  const caller = ({ video }) => {
+    const { _id, avatar, username, fullname } = user;
+    const msg = {
+      sender: auth.user._id,
+      recipient: _id,
+      avatar,
+      username,
+      fullname,
+      video,
+    };
+    dispatch({ type: TYPES.CALL, payload: msg });
+  };
+
+  const handleAudioCall = () => {
+    caller({ video: false });
+  };
+
+  const handleVideoCall = () => {
+    caller({ video: true });
   };
   //
   return (
@@ -162,10 +187,15 @@ const RightSide = () => {
       <div className="message_header" style={{ cursor: 'pointer' }}>
         {user.length !== 0 && (
           <UserCard user={user}>
-            <i
-              className="fas fa-trash text-danger"
-              onClick={handleDelteConversation}
-            />
+            <div>
+              <i className="fas fa-phone-alt" onClick={handleAudioCall} />
+
+              <i className="fas fa-video mx-3" onClick={handleVideoCall} />
+              <i
+                className="fas fa-trash text-danger"
+                onClick={handleDelteConversation}
+              />
+            </div>
           </UserCard>
         )}
       </div>
